Corrected problem in stf, and more than likely elsewher if you try and copy between objects with different dimensions. For example
x[,1:3,,,,1:100]<-y[,1,,,,1:100]

In this case you will copy iter 1 into year 1,
iter 2 into year 2
and
iter 3 into year 3

You might not realise this is happening and if iters are random noise it might not matter. However if the other dims don´t match then you are well and truely screwed!